+++ lib/SKGB/Intern/Controller/Content.pm
@@ -31,7 +31,7 @@ sub stegdienstliste {
 	
 	my @records = $self->neo4j->get_persons(<<END, column => 'p');
 MATCH (p:Person)-[r:ROLE|GUEST]->(Role{role:'active-member'})
-WHERE NOT(has(r.leaves))
+WHERE NOT(exists(r.leaves))
 OPTIONAL MATCH (p)<-[:FOR]-(a:Address{type:'street'})
 OPTIONAL MATCH (p)-[:ROLE|GUEST*..3]->(v:Role{role:'board-member'})
 RETURN [p, r] as p, a.place AS place, v AS board, r.noService as exempt